+++ packages/atomic/cypress/e2e/facets/rating-range-facet/rating-range-facet-selectors.ts
@@ -1,3 +1,5 @@
+import {TestFixture} from '../../../fixtures/test-fixture';
+
 export const ratingRangeFacetComponent = 'atomic-rating-range-facet';
 export const RatingRangeFacetSelectors = {
   withId(id: string) {
@@ -61,6 +63,8 @@ export const RatingRangeFacetSelectors = {
     return this.valueLabel().eq(index);
   },
   starsIconAtIndex(index: number) {
-    return this.facetValueAtIndex(index).find('atomic-icon');
+    return cy
+      .wait(TestFixture.interceptAliases.Build)
+      .then(() => this.facetValueAtIndex(index).find('atomic-icon'));
   },
 };