+++ packages/atomic/src/components/search/facets/atomic-color-facet/atomic-color-facet.tsx
@@ -132,7 +132,7 @@ export class AtomicColorFacet
   @Prop({reflect: true}) public numberOfValues = 8;
   /**
    * Whether this facet should contain a search box.
-   * When "true", the search is only enabled when more facet values are available.
+   *
    */
   @Prop({reflect: true}) public withSearch = true;
   /**