+++ packages/atomic/src/components/search/formats/atomic-format-unit.numeric-facet.stories.tsx
@@ -19,4 +19,4 @@ export default {
   ...defaultModuleExport,
   title: 'Atomic/NumericFacet/Format/Unit',
 };
-export const DefaultFormatNumber = exportedStory;
+export const Default = exportedStory;