+++ packages/atomic/src/components/search/result-template-components/atomic-result-link/atomic-result-link.stories.tsx
@@ -7,4 +7,4 @@ const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
 );
 
 export default {...defaultModuleExport, title: 'Atomic/ResultList/ResultLink'};
-export const DefaultResultLink = exportedStory;
+export const Default = exportedStory;