+++ packages/headless/src/controllers/generated-answer/headless-generated-answer.test.ts
@@ -101,6 +101,17 @@ describe('generated answer', () => {
     expect(action).toBeTruthy();
     expect(analyticsAction).toBeTruthy();
   });
+  it('#like dispatches no analytics action when #liked is set to true', () => {
+    engine.state.generatedAnswer.liked = true;
+    generatedAnswer.like();
+    const action = findAction(likeGeneratedAnswer.type);
+    const analyticsAction = engine.findAsyncAction(
+      logLikeGeneratedAnswer().pending
+    );
+
+    expect(action).not.toBeTruthy();
+    expect(analyticsAction).not.toBeTruthy();
+  });
 
   it('#dislike dispatches analytics action', () => {
     generatedAnswer.dislike();
@@ -111,6 +122,18 @@ describe('generated answer', () => {
 
     expect(action).toBeTruthy();
     expect(analyticsAction).toBeTruthy();
+  });
+
+  it('#dislike dispatches no analytics action when #disliked is set to true', () => {
+    engine.state.generatedAnswer.disliked = true;
+    generatedAnswer.dislike();
+    const action = findAction(dislikeGeneratedAnswer.type);
+    const analyticsAction = engine.findAsyncAction(
+      logDislikeGeneratedAnswer().pending
+    );
+
+    expect(action).not.toBeTruthy();
+    expect(analyticsAction).not.toBeTruthy();
   });
 
   it('#openFeedbackModal dispatches the right actions', () => {
@@ -241,9 +264,11 @@ describe('generated answer', () => {
 
   describe('subscription to changes', () => {
     function callListener() {
-      return (engine.subscribe as jest.Mock).mock.calls.map(
-        (args) => args[0]
-      )[0]();
+      return (engine.subscribe as jest.Mock).mock.calls
+        .map((args) => args[0])
+        .forEach((listener) => {
+          listener();
+        });
     }
 
     it('should not dispatch the stream action when there is no stream ID', () => {
@@ -274,6 +299,28 @@ describe('generated answer', () => {
       const action = findAction(streamAnswer.pending.type);
 
       expect(action).toBeTruthy();
+    });
+
+    describe('when we have multiple controllers', () => {
+      let secondEngine: MockSearchEngine;
+
+      beforeEach(() => {
+        secondEngine = buildMockSearchAppEngine();
+        buildGeneratedAnswer(secondEngine);
+      });
+
+      it('should dispatch the stream action only once', () => {
+        engine.state.search.extendedResults = {
+          generativeQuestionAnsweringId: 'another-fake-test-id',
+        };
+
+        callListener();
+
+        const count = engine.actions.filter(
+          (a) => a.type === streamAnswer.pending.type
+        ).length;
+        expect(count).toEqual(1);
+      });
     });
   });