+++ packages/quantic/force-app/main/default/lwc/quanticResultLink/__tests__/quanticResultLink.test.js
@@ -32,7 +32,7 @@ function flushPromises() {
 }
 
 // Helper function to mock headless for this suite of tests.
-function mockHeadless() {
+function mockHeadlessAndBueno() {
   jest.spyOn(mockHeadlessLoader, 'getHeadlessBundle').mockReturnValue({
     buildInteractiveResult: () => ({
       select: jest.fn(),
@@ -44,6 +44,20 @@ function mockHeadless() {
   jest.spyOn(mockHeadlessLoader, 'getHeadlessEnginePromise').mockReturnValue(
     new Promise((resolve) => {
       resolve();
+    })
+  );
+
+  jest.spyOn(mockHeadlessLoader, 'getBueno').mockReturnValue(
+    new Promise(() => {
+      // @ts-ignore
+      global.Bueno = {
+        isString: jest
+          .fn()
+          .mockImplementation(
+            (value) =>
+              Object.prototype.toString.call(value) === '[object String]'
+          ),
+      };
     })
   );
 }
@@ -58,7 +72,7 @@ describe('c-quantic-result-link', () => {
   }
 
   beforeEach(() => {
-    mockHeadless();
+    mockHeadlessAndBueno();
   });
 
   afterEach(() => {