+++ packages/quantic/force-app/main/default/lwc/quanticSourceCitations/quanticSourceCitations.js
@@ -1,4 +1,4 @@
-import learnMore from '@salesforce/label/c.quantic_LearnMore';
+import citations from '@salesforce/label/c.quantic_Citations';
 import {LightningElement, api} from 'lwc';
 
 /** @typedef {import("coveo").GeneratedAnswerCitation} GeneratedAnswerCitation */
@@ -11,7 +11,7 @@ import {LightningElement, api} from 'lwc';
  */
 export default class QuanticSourceCitations extends LightningElement {
   labels = {
-    learnMore,
+    citations,
   };
 
   /**
@@ -26,6 +26,12 @@ export default class QuanticSourceCitations extends LightningElement {
    * @type {function}
    */
   @api citationClickHandler;
+  /**
+   * The function to be executed when a citation is clicked.
+   * @api
+   * @type {function}
+   */
+  @api citationHoverHandler;
 
   /**
    * Returns the indexed citations.
@@ -33,10 +39,18 @@ export default class QuanticSourceCitations extends LightningElement {
    */
   get indexedCitations() {
     return this.citations.map((citation, index) => ({
-      ...citation,
-      index: index + 1,
+      data: {
+        ...citation,
+        index: index + 1,
+      },
       handleCitationClick: () => {
         this.citationClickHandler?.(citation.id);
+      },
+      handleCitationHover: (event) => {
+        this.citationHoverHandler?.(
+          citation.id,
+          event.detail.citationHoverTimeMs
+        );
       },
     }));
   }