+++ packages/samples/headless-ssr/common/components/generic/search-page.tsx
@@ -22,11 +22,15 @@ export default function SearchPage({
   >(undefined);
 
   useEffect(() => {
+    const {urlManager, context} = staticState.controllers;
     hydrateStaticState({
       searchAction: staticState.searchAction,
       controllers: {
+        context: {
+          initialState: context.state,
+        },
         urlManager: {
-          initialState: staticState.controllers.urlManager.state,
+          initialState: urlManager.state,
         },
       },
     }).then(({engine, controllers}) => {